+++ eelbrain/_utils/__init__.py
@@ -1,5 +1,4 @@
-from .basic import (IdDict, intervals, LazyProperty, keydefaultdict, natsorted,
-                    set_log_level)
-
+from .basic import (deprecated, IdDict, intervals, LazyProperty, keydefaultdict,
+                    natsorted, set_log_level)
 from . import kit
 from .system import caffeine